test(depeg): retire the DDR historical replay corpus and merge the pending suites
The DDR test surface carried a replay corpus of 26 migration fixtures for
one-off historical data repairs (0135-0227: sealed-reopen links, tail
relinks, duplicate-prediction errata, per-event splits). Those migrations
are squashed into the base schema, so the fixtures only existed to replay
history that can no longer occur, and `latest-schema-sqlite` had to walk
all of them to build a test schema.

Also merges the two single-case satellite suites
(confirm-pending-depegs-opposite-direction, -pool-status) into
confirm-pending-depegs.test.ts, which already owned the harness they
each rebuilt, and shares the public-projection fixtures through a new
depeg-public-projection.test-support.ts.

Case accounting: the two satellites contributed 1 case each and
confirm-pending-depegs.test.ts went 35 -> 37, so the merge is
case-preserving. Separately, 5 cases were removed from
depeg-resolver-ddrv2-store.test.ts because they asserted the behaviour of
specific retired migrations and cannot run without that corpus:

  - invalidates sealed APXUSD duplicate predictions without relink order hazards
  - keeps DDR repair-debt cache markers that include unrelated cache-only debt
  - keeps cache-only DDR repair-debt markers when event 90410 is not repaired
  - ledgers LUSD event 90410 as a fresh DDR incident and closes repair debt
  - repairs APXUSD event 90203 when the earlier relink migration found an accidental link

Those five pinned immutable past behaviour: the migrations already ran in
production and cannot regress. Every live DDR contract keeps its
coverage. This is the only intentional coverage reduction in the wave and
is called out here so it stays reviewable.

Net -9,818 lines.
