+++ worker/src/cron/__tests__/depeg-resolver-public-projection.test.ts
@@ -21,6 +21,7 @@ import {
   buildV2PublicationBasePayload,
   normalizeErratumRecord,
 } from "../depeg-resolver/public-projection";
+import { makeDdrCanonicalIncident, makeDdrResolverRow } from "./depeg-public-projection.test-support";
 
 describe("depeg-resolver public projection", () => {
   const NOW_SEC = 1_779_984_600;
@@ -39,64 +40,7 @@ describe("depeg-resolver public projection", () => {
   }
 
   function resolverRow(overrides: Partial<DdrRow> = {}): DdrRow {
-    const startedAt = overrides.startedAt ?? NOW_SEC - 6 * 3600;
-    return {
-      stablecoinId: "projection-test",
-      symbol: "PRJ",
-      name: "Projection Test",
-      pegCurrency: "USD",
-      governance: "decentralized",
-      status: "active",
-      eventId: 42,
-      startedAt,
-      ageSec: NOW_SEC - startedAt,
-      direction: "below",
-      peakDeviationBps: -250,
-      currentDeviationBps: -180,
-      resolution: {
-        tier: "recovery_likely",
-        factors: [
-          {
-            code: "R2_hard_collateral_redemption",
-            kind: "anchor",
-            severity: "strong",
-            label: "Fixture has a hard collateral recovery anchor",
-          },
-        ],
-      },
-      duration: {
-        suppressed: false,
-        suppressedReason: null,
-        stratum: "below - moderate - robust - USD",
-        medianSec: 3600,
-        iqrSec: [1800, 7200],
-        ageStatus: "ordinary",
-        horizons: [
-          {
-            horizon: "6h",
-            state: "benchmarked",
-            probability: 0.75,
-            probabilityDisplay: "70-80%",
-            probabilityInterval: { lower: 0.7, upper: 0.8 },
-            rawAtRisk: 20,
-            uniqueCoins: 12,
-            intervalClosures: 15,
-            intervalNonClosures: 5,
-          },
-        ],
-      },
-      relatedContext: {
-        dewsBand: "WATCH",
-        dewsScore: 22,
-        liquidityScore: 88,
-        safetyGrade: "A",
-        safetyScore: 91,
-        supplyChange7dPct: 0,
-        supplyChange30dPct: 0,
-        mintSurge: false,
-      },
-      ...overrides,
-    };
+    return makeDdrResolverRow(overrides, { nowSec: NOW_SEC });
   }
 
   function insufficientSignalRow(overrides: Partial<DdrRow> = {}): DdrRow {
@@ -121,21 +65,7 @@ describe("depeg-resolver public projection", () => {
   }
 
   function canonicalIncident(row: DdrRow, overrides: Partial<DdrCanonicalIncident> = {}): DdrCanonicalIncident {
-    return {
-      incidentKey: `ddr2:projection-${row.eventId}`,
-      eventId: row.eventId,
-      currentEventId: row.eventId,
-      stablecoinId: row.stablecoinId,
-      pegCurrency: row.pegCurrency,
-      direction: row.direction,
-      startedAt: row.startedAt,
-      eligibleAt: row.startedAt + DAY_SEC,
-      policyUniverseIncluded: true,
-      rolloutActiveAtEnablement: true,
-      confirmedAt: null,
-      lockState: null,
-      ...overrides,
-    };
+    return makeDdrCanonicalIncident(row, { eligibleAt: row.startedAt + DAY_SEC, ...overrides });
   }
 
   function sealedPrediction(
