+++ worker/src/lib/__tests__/depeg-resolver-ddrv2-store.test.ts
@@ -2,11 +2,7 @@ import { DatabaseSync } from "node:sqlite";
 import type { D1Database } from "@cloudflare/workers-types";
 import { describe, expect, it } from "vitest";
 import { mockD1 } from "../../test-helpers/__shared/mock-d1";
-import {
-  applyMigrationFile,
-  applyMigrationsThrough,
-  createLatestSchemaSqlite,
-} from "../../test-helpers/latest-schema-sqlite";
+import { createLatestSchemaSqlite } from "../../test-helpers/latest-schema-sqlite";
 import {
   closeRecoveredPreLockIncidents,
   DDR_FLAP_TOLERANT_MAX_INCIDENT_SPAN_SEC_V1,
@@ -515,477 +511,6 @@ async function seedFlapMigrationIncident(
 }
 
 describe("DDRv2 storage migrations and stores", () => {
-  it("ledgers LUSD event 90410 as a fresh DDR incident and closes repair debt", () => {
-    const sqlite = new DatabaseSync(":memory:");
-    try {
-      applyMigrationsThrough(sqlite, "0168_surface_publication_generations.sql");
-
-      sqlite.exec(`
-        INSERT INTO depeg_events
-          (id, stablecoin_id, symbol, peg_type, direction, peak_deviation_bps,
-           started_at, ended_at, start_price, peak_price, recovery_price, peg_reference, source)
-        VALUES
-          (90404, 'lusd-liquity', 'LUSD', 'peggedUSD', 'above', 100,
-           1782944363, 1782947931, 1.0096705168234137, 1.0096705168234137, 1.0085284741080422, 0.9997113947811964, 'live'),
-          (90410, 'lusd-liquity', 'LUSD', 'peggedUSD', 'above', 100,
-           1782970434, 1782973127, 1.0096702857093371, 1.0096702857093371, 1.0092109177692081, 0.99968, 'live');
-
-        INSERT INTO depeg_resolver_incident_event_links
-          (incident_key, event_id, relation, repair_authorization_id, linked_at, note)
-        VALUES
-          ('ddr2:ba7b7cb3eb0ef0237e1fac09475cd198', 90404, 'repair_replacement', NULL, 1782944403,
-           'pre-lock nearby event adopted as current incident source');
-
-        INSERT INTO depeg_resolver_incidents
-          (incident_key, stablecoin_id, peg_currency, direction, first_event_id, current_event_id,
-           first_started_at, current_started_at, first_observed_peak_bucket_bps, incident_state,
-           superseded_by_incident_key, source_fingerprint, created_at, updated_at)
-        VALUES
-          ('ddr2:ba7b7cb3eb0ef0237e1fac09475cd198', 'lusd-liquity', 'USD', 'above',
-           90361, 90404, 1782702288, 1782944363, 100, 'active', NULL,
-           'a812e8287ccbedb13ec00d3b4d8a2ec1501b85dbf131a332a14715a6836c360b', 1782702297, 1782944403);
-
-        INSERT INTO worker_repair_tasks
-          (task_id, kind, subject_id, priority, state, attempt_count, next_attempt_at,
-           payload_json, created_at, updated_at)
-        VALUES
-          ('repair:ddr-repair-required-event:90410', 'ddr-repair-required-event', '90410', 50, 'open', 0, NULL,
-           '{"eventId":90410,"reason":"Unlinked depeg event 90410 overlaps nearby canonical incident ddr2:ba7b7cb3eb0ef0237e1fac09475cd198; explicit repair required"}',
-           1782970440, 1782973133);
-
-        INSERT INTO cache (key, value, updated_at)
-        VALUES (
-          'ddr:repair-debt:v1',
-          '{"checkedAt":1782973133,"count":1,"events":[{"eventId":90410,"reason":"Unlinked depeg event 90410 overlaps nearby canonical incident ddr2:ba7b7cb3eb0ef0237e1fac09475cd198; explicit repair required"}],"eventsTruncated":false}',
-          1782973133
-        );
-      `);
-
-      applyMigrationFile(sqlite, "0169_lusd_ddr_event_90410_split.sql");
-
-      expect(
-        sqlite
-          .prepare(
-            `SELECT incident_key, relation, note
-             FROM depeg_resolver_incident_event_links
-             WHERE event_id = 90410`,
-          )
-          .get(),
-      ).toEqual({
-        incident_key: "ddr2:c14884852abe024faa4d4b9fc1f84742",
-        relation: "observed",
-        note: "fresh LUSD flap split from earlier recovered pre-lock chain",
-      });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT first_event_id, current_event_id, source_fingerprint
-             FROM depeg_resolver_incidents
-             WHERE incident_key = 'ddr2:c14884852abe024faa4d4b9fc1f84742'`,
-          )
-          .get(),
-      ).toEqual({
-        first_event_id: 90410,
-        current_event_id: 90410,
-        source_fingerprint: "09da1784f586d5bbfba3762fa54d4c3c1d47f1660fd5c315616a3187b480f687",
-      });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT policy_universe_included, policy_universe_reason
-             FROM depeg_resolver_incident_policy_membership
-             WHERE incident_key = 'ddr2:c14884852abe024faa4d4b9fc1f84742'`,
-          )
-          .get(),
-      ).toEqual({
-        policy_universe_included: 1,
-        policy_universe_reason: "post_effective_public_tracked",
-      });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT current_event_id
-             FROM depeg_resolver_incidents
-             WHERE incident_key = 'ddr2:ba7b7cb3eb0ef0237e1fac09475cd198'`,
-          )
-          .get(),
-      ).toEqual({ current_event_id: 90404 });
-      expect(
-        sqlite
-          .prepare("SELECT state, closed_at IS NOT NULL AS has_closed_at FROM worker_repair_tasks WHERE subject_id = '90410'")
-          .get(),
-      ).toEqual({ state: "closed", has_closed_at: 1 });
-      expect(
-        sqlite
-          .prepare("SELECT COUNT(*) AS count FROM cache WHERE key = 'ddr:repair-debt:v1'")
-          .get(),
-      ).toEqual({ count: 0 });
-    } finally {
-      sqlite.close();
-    }
-  });
-
-  it("keeps DDR repair-debt cache markers that include unrelated cache-only debt", () => {
-    const sqlite = new DatabaseSync(":memory:");
-    try {
-      applyMigrationsThrough(sqlite, "0168_surface_publication_generations.sql");
-
-      sqlite.exec(`
-        INSERT INTO depeg_events
-          (id, stablecoin_id, symbol, peg_type, direction, peak_deviation_bps,
-           started_at, ended_at, start_price, peak_price, recovery_price, peg_reference, source)
-        VALUES
-          (90410, 'lusd-liquity', 'LUSD', 'peggedUSD', 'above', 100,
-           1782970434, 1782973127, 1.0096702857093371, 1.0096702857093371, 1.0092109177692081, 0.99968, 'live');
-
-        INSERT INTO worker_repair_tasks
-          (task_id, kind, subject_id, priority, state, attempt_count, next_attempt_at,
-           payload_json, created_at, updated_at)
-        VALUES
-          ('repair:ddr-repair-required-event:90410', 'ddr-repair-required-event', '90410', 50, 'open', 0, NULL,
-           '{"eventId":90410,"reason":"explicit repair required"}',
-           1782970440, 1782973133);
-
-        INSERT INTO cache (key, value, updated_at)
-        VALUES (
-          'ddr:repair-debt:v1',
-          '{"checkedAt":1782973133,"count":2,"events":[{"eventId":90410,"reason":"explicit repair required"},{"eventId":99999,"reason":"cache-only repair debt"}],"eventsTruncated":false}',
-          1782973133
-        );
-      `);
-
-      applyMigrationFile(sqlite, "0169_lusd_ddr_event_90410_split.sql");
-
-      expect(
-        sqlite
-          .prepare("SELECT state, closed_at IS NOT NULL AS has_closed_at FROM worker_repair_tasks WHERE subject_id = '90410'")
-          .get(),
-      ).toEqual({ state: "closed", has_closed_at: 1 });
-      expect(
-        sqlite
-          .prepare("SELECT value FROM cache WHERE key = 'ddr:repair-debt:v1'")
-          .get(),
-      ).toEqual({
-        value:
-          '{"checkedAt":1782973133,"count":2,"events":[{"eventId":90410,"reason":"explicit repair required"},{"eventId":99999,"reason":"cache-only repair debt"}],"eventsTruncated":false}',
-      });
-    } finally {
-      sqlite.close();
-    }
-  });
-
-  it("keeps cache-only DDR repair-debt markers when event 90410 is not repaired", () => {
-    const sqlite = new DatabaseSync(":memory:");
-    try {
-      applyMigrationsThrough(sqlite, "0168_surface_publication_generations.sql");
-
-      sqlite.exec(`
-        INSERT INTO cache (key, value, updated_at)
-        VALUES (
-          'ddr:repair-debt:v1',
-          '{"checkedAt":1782973133,"count":1,"events":[{"eventId":99999,"reason":"cache-only repair debt"}],"eventsTruncated":false}',
-          1782973133
-        );
-      `);
-
-      applyMigrationFile(sqlite, "0169_lusd_ddr_event_90410_split.sql");
-
-      expect(
-        sqlite
-          .prepare("SELECT value FROM cache WHERE key = 'ddr:repair-debt:v1'")
-          .get(),
-      ).toEqual({
-        value: '{"checkedAt":1782973133,"count":1,"events":[{"eventId":99999,"reason":"cache-only repair debt"}],"eventsTruncated":false}',
-      });
-    } finally {
-      sqlite.close();
-    }
-  });
-
-  it("repairs APXUSD event 90203 when the earlier relink migration found an accidental link", () => {
-    const sqlite = new DatabaseSync(":memory:");
-    try {
-      applyMigrationsThrough(sqlite, "0153_status_reliability_idempotency.sql");
-
-      sqlite.exec(`
-        INSERT INTO depeg_events
-          (id, stablecoin_id, symbol, peg_type, direction, peak_deviation_bps,
-           started_at, ended_at, start_price, peak_price, recovery_price, peg_reference, source)
-        VALUES
-          (90089, 'apxusd-apyx', 'apxUSD', 'peggedUSD', 'below', -1200,
-           1780671044, 1781624981, 0.99, 0.88, 1.00, 1, 'live'),
-          (90203, 'apxusd-apyx', 'apxUSD', 'peggedUSD', 'below', -900,
-           1781632159, NULL, 0.99, 0.91, NULL, 1, 'live');
-
-        INSERT INTO depeg_resolver_incident_event_links
-          (incident_key, event_id, relation, repair_authorization_id, linked_at, note)
-        VALUES
-          ('ddr2:e32c8186781838eac1b740a44c3b8776', 90089, 'observed', NULL, 1781625000, 'canonical event'),
-          ('ddr2:70a8e43c093e0afcdc8a37143a6849f9', 90203, 'observed', NULL, 1781632200, 'accidental tail');
-
-        INSERT INTO depeg_resolver_incidents
-          (incident_key, stablecoin_id, peg_currency, direction, first_event_id, current_event_id,
-           first_started_at, current_started_at, first_observed_peak_bucket_bps, incident_state,
-           superseded_by_incident_key, source_fingerprint, created_at, updated_at)
-        VALUES
-          ('ddr2:e32c8186781838eac1b740a44c3b8776', 'apxusd-apyx', 'USD', 'below',
-           90089, 90089, 1780671044, 1780671044, 1200, 'active', NULL,
-           'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa', 1781625000, 1781625000),
-          ('ddr2:70a8e43c093e0afcdc8a37143a6849f9', 'apxusd-apyx', 'USD', 'below',
-           90203, 90203, 1781632159, 1781632159, 900, 'active', NULL,
-           'bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb', 1781632200, 1781632200);
-
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_relational_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_version_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_payload_identity_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_payload_prediction_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_prediction_kind_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_no_call_kind_guard;
-
-        INSERT INTO depeg_resolver_public_predictions
-          (incident_key, event_id, assessment_id, outcome_kind, prediction_policy_version,
-           prediction_methodology_version, prediction_methodology_version_label, resolution_rubric_version,
-           duration_model_version, incident_grouping_version, support_rules_version, policy_delay_sec,
-           eligible_at, locked_at, event_age_at_lock_sec, lock_timing, sealed_payload_json, row_hash, created_at)
-        VALUES
-          ('ddr2:e32c8186781838eac1b740a44c3b8776', 90089, 1, 'prediction', 'sticky-24h-v1',
-           'v1', 'v1', 'rubric-v1', 'duration-v1', 'group-v1', 'support-v1', 86400,
-           1780757444, 1780757444, 86400, 'on_time', '{}',
-           'cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc', 1780757444);
-
-        INSERT INTO cache (key, value, updated_at)
-        VALUES
-          ('depeg-resolver:snapshot', '{}', 1781632300),
-          ('depeg-resolver-review:snapshot', '{}', 1781632300);
-      `);
-
-      applyMigrationFile(sqlite, "0154_apxusd_ddr_tail_90203_link.sql");
-
-      expect(
-        sqlite
-          .prepare("SELECT incident_key FROM depeg_resolver_incident_event_links WHERE event_id = 90203")
-          .get(),
-      ).toEqual({ incident_key: "ddr2:70a8e43c093e0afcdc8a37143a6849f9" });
-      expect(
-        sqlite
-          .prepare("SELECT current_event_id FROM depeg_resolver_incidents WHERE incident_key = ?")
-          .get("ddr2:e32c8186781838eac1b740a44c3b8776"),
-      ).toEqual({ current_event_id: 90089 });
-
-      applyMigrationFile(sqlite, "0161_apxusd_ddr_tail_90203_relink_repair.sql");
-
-      expect(
-        sqlite
-          .prepare(
-            `SELECT incident_key, relation
-             FROM depeg_resolver_incident_event_links
-             WHERE event_id = 90203`,
-          )
-          .get(),
-      ).toEqual({
-        incident_key: "ddr2:e32c8186781838eac1b740a44c3b8776",
-        relation: "repair_replacement",
-      });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT current_event_id
-             FROM depeg_resolver_incidents
-             WHERE incident_key = ?`,
-          )
-          .get("ddr2:e32c8186781838eac1b740a44c3b8776"),
-      ).toEqual({ current_event_id: 90203 });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT incident_state, superseded_by_incident_key
-             FROM depeg_resolver_incidents
-             WHERE incident_key = ?`,
-          )
-          .get("ddr2:70a8e43c093e0afcdc8a37143a6849f9"),
-      ).toEqual({
-        incident_state: "superseded",
-        superseded_by_incident_key: "ddr2:e32c8186781838eac1b740a44c3b8776",
-      });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT COUNT(*) AS count
-             FROM depeg_resolver_event_repair_authorization_uses u
-             JOIN depeg_resolver_event_repair_authorizations a
-               ON a.id = u.authorization_id
-             WHERE a.event_id = 90203
-               AND a.incident_key = 'ddr2:e32c8186781838eac1b740a44c3b8776'
-               AND a.operation IN ('incident_link', 'incident_current_update')`,
-          )
-          .get(),
-      ).toEqual({ count: 2 });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT sql
-             FROM sqlite_master
-             WHERE type = 'index'
-               AND name = 'idx_ddr_incident_current_event'`,
-          )
-          .get(),
-      ).toEqual({
-        sql: "CREATE UNIQUE INDEX idx_ddr_incident_current_event\n  ON depeg_resolver_incidents(current_event_id)\n  WHERE incident_state = 'active'",
-      });
-    } finally {
-      sqlite.close();
-    }
-  });
-
-  it("invalidates sealed APXUSD duplicate predictions without relink order hazards", () => {
-    const sqlite = new DatabaseSync(":memory:");
-    try {
-      applyMigrationsThrough(sqlite, "0153_status_reliability_idempotency.sql");
-
-      sqlite.exec(`
-        INSERT INTO depeg_events
-          (id, stablecoin_id, symbol, peg_type, direction, peak_deviation_bps,
-           started_at, ended_at, start_price, peak_price, recovery_price, peg_reference, source)
-        VALUES
-          (90089, 'apxusd-apyx', 'apxUSD', 'peggedUSD', 'below', -1200,
-           1780671044, 1781624981, 0.99, 0.88, 1.00, 1, 'live'),
-          (90203, 'apxusd-apyx', 'apxUSD', 'peggedUSD', 'below', -900,
-           1781632159, NULL, 0.99, 0.91, NULL, 1, 'live');
-
-        INSERT INTO depeg_resolver_incident_event_links
-          (incident_key, event_id, relation, repair_authorization_id, linked_at, note)
-        VALUES
-          ('ddr2:e32c8186781838eac1b740a44c3b8776', 90089, 'observed', NULL, 1781625000, 'canonical event'),
-          ('ddr2:70a8e43c093e0afcdc8a37143a6849f9', 90203, 'observed', NULL, 1781632200, 'accidental tail');
-
-        INSERT INTO depeg_resolver_incidents
-          (incident_key, stablecoin_id, peg_currency, direction, first_event_id, current_event_id,
-           first_started_at, current_started_at, first_observed_peak_bucket_bps, incident_state,
-           superseded_by_incident_key, source_fingerprint, created_at, updated_at)
-        VALUES
-          ('ddr2:e32c8186781838eac1b740a44c3b8776', 'apxusd-apyx', 'USD', 'below',
-           90089, 90089, 1780671044, 1780671044, 1200, 'active', NULL,
-           'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa', 1781625000, 1781625000),
-          ('ddr2:70a8e43c093e0afcdc8a37143a6849f9', 'apxusd-apyx', 'USD', 'below',
-           90203, 90203, 1781632159, 1781632159, 900, 'active', NULL,
-           'bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb', 1781632200, 1781632200);
-
-        INSERT INTO depeg_resolver_assessments
-          (id, event_id, stablecoin_id, symbol, name, peg_currency, governance, direction,
-           started_at, assessed_at, event_age_sec, checkpoint, methodology_version,
-           methodology_version_label, resolution_rubric_version, duration_model_version,
-           incident_grouping_version, support_rules_version, resolution_tier,
-           duration_suppressed, duration_suppressed_reason, median_remaining_sec,
-           iqr_low_remaining_sec, iqr_high_remaining_sec, stratum, horizons_json,
-           factors_json, row_json, created_at, updated_at)
-        VALUES
-          (1, 90089, 'apxusd-apyx', 'apxUSD', 'APX USD', 'USD', 'centralized', 'below',
-           1780671044, 1780757444, 86400, 'public_prediction', 'v1', 'v1',
-           'rubric-v1', 'duration-v1', 'group-v1', 'support-v1', 'at_risk',
-           0, NULL, NULL, NULL, NULL, NULL, '[]', '[]', '{}', 1780757444, 1780757444),
-          (2, 90203, 'apxusd-apyx', 'apxUSD', 'APX USD', 'USD', 'centralized', 'below',
-           1781632159, 1781718559, 86400, 'public_prediction', 'v1', 'v1',
-           'rubric-v1', 'duration-v1', 'group-v1', 'support-v1', 'at_risk',
-           0, NULL, NULL, NULL, NULL, NULL, '[]', '[]', '{}', 1781718559, 1781718559);
-
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_relational_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_version_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_payload_identity_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_payload_prediction_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_prediction_kind_guard;
-        DROP TRIGGER IF EXISTS trg_ddr_public_predictions_no_call_kind_guard;
-
-        INSERT INTO depeg_resolver_public_predictions
-          (incident_key, event_id, assessment_id, outcome_kind, prediction_policy_version,
-           prediction_methodology_version, prediction_methodology_version_label, resolution_rubric_version,
-           duration_model_version, incident_grouping_version, support_rules_version, policy_delay_sec,
-           eligible_at, locked_at, event_age_at_lock_sec, lock_timing, sealed_payload_json, row_hash, created_at)
-        VALUES
-          ('ddr2:e32c8186781838eac1b740a44c3b8776', 90089, 1, 'prediction', 'sticky-24h-v1',
-           'v1', 'v1', 'rubric-v1', 'duration-v1', 'group-v1', 'support-v1', 86400,
-           1780757444, 1780757444, 86400, 'on_time', '{}',
-           'cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc', 1780757444),
-          ('ddr2:70a8e43c093e0afcdc8a37143a6849f9', 90203, 2, 'prediction', 'sticky-24h-v1',
-           'v1', 'v1', 'rubric-v1', 'duration-v1', 'group-v1', 'support-v1', 86400,
-           1781718559, 1781718559, 86400, 'on_time', '{}',
-           'dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd', 1781718559);
-
-        INSERT INTO cache (key, value, updated_at)
-        VALUES
-          ('depeg-resolver:snapshot', '{}', 1781632300),
-          ('depeg-resolver-review:snapshot', '{}', 1781632300);
-      `);
-
-      applyMigrationFile(sqlite, "0154_apxusd_ddr_tail_90203_link.sql");
-      applyMigrationFile(sqlite, "0161_apxusd_ddr_tail_90203_relink_repair.sql");
-
-      expect(
-        sqlite
-          .prepare("SELECT incident_key FROM depeg_resolver_incident_event_links WHERE event_id = 90203")
-          .get(),
-      ).toEqual({ incident_key: "ddr2:70a8e43c093e0afcdc8a37143a6849f9" });
-      expect(
-        sqlite
-          .prepare("SELECT current_event_id FROM depeg_resolver_incidents WHERE incident_key = ?")
-          .get("ddr2:e32c8186781838eac1b740a44c3b8776"),
-      ).toEqual({ current_event_id: 90089 });
-
-      applyMigrationFile(sqlite, "0162_apxusd_duplicate_ddr_prediction_erratum.sql");
-
-      expect(
-        sqlite
-          .prepare(
-            `SELECT incident_state, superseded_by_incident_key
-             FROM depeg_resolver_incidents
-             WHERE incident_key = ?`,
-          )
-          .get("ddr2:70a8e43c093e0afcdc8a37143a6849f9"),
-      ).toEqual({
-        incident_state: "superseded",
-        superseded_by_incident_key: "ddr2:e32c8186781838eac1b740a44c3b8776",
-      });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT e.reason, e.replacement_assessment_id, e.created_by
-             FROM depeg_resolver_prediction_errata e
-             JOIN depeg_resolver_public_predictions p ON p.id = e.public_prediction_id
-             WHERE p.incident_key = 'ddr2:70a8e43c093e0afcdc8a37143a6849f9'`,
-          )
-          .get(),
-      ).toEqual({
-        reason: "event_identity_error",
-        replacement_assessment_id: 1,
-        created_by: "migration-0162",
-      });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT relation, created_by
-             FROM depeg_resolver_incident_lineage
-             WHERE from_incident_key = 'ddr2:70a8e43c093e0afcdc8a37143a6849f9'
-               AND to_incident_key = 'ddr2:e32c8186781838eac1b740a44c3b8776'`,
-          )
-          .get(),
-      ).toEqual({ relation: "superseded_by", created_by: "migration-0162" });
-      expect(
-        sqlite
-          .prepare(
-            `SELECT COUNT(*) AS count
-             FROM depeg_resolver_event_repair_authorization_uses u
-             JOIN depeg_resolver_event_repair_authorizations a
-               ON a.id = u.authorization_id
-             WHERE a.event_id = 90203
-               AND a.incident_key = 'ddr2:70a8e43c093e0afcdc8a37143a6849f9'
-               AND a.operation = 'incident_current_update'`,
-          )
-          .get(),
-      ).toEqual({ count: 1 });
-    } finally {
-      sqlite.close();
-    }
-  });
-
   it("replaces the monolithic public-prediction guard with split triggers", () => {
     const db = makeSqliteD1();
     try {
